pyproject.toml
[project]
name = "auto-icd-o"
version = "0.1.0"
description = ""
authors = [
    {name = "Kamyar Arzideh",email = "kamyar.arzideh@uk-essen.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "openai (>=2.6.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "wandb (>=0.22.2,<0.23.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "adjusttext (>=1.3.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "auto_icd_o", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
